requirements.txt
streamlit
langchain
langchain-openai
langchain-community 
azure-search-documents
azure-identity
pypdf
python-dotenv